refactor: createLoggerを移動
